+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.13.0 - In Development
 
 * Added plugin system. See the plugins page in the project documentation for details.
+* Added automatic session creation. Added new `--password` option to specify the password to use to login.
 * Added collection of potential injection points. These are available to plugins via a hook or JSON output.
 * Reduced memory usage by approx. 30% + approx. 275MB.
 * Added support for configuration settings in `~/.yawast-ng.json` (see Usage page in the documentation).